:recycle: Refactor: controller test 코드 위치 이동
<footer>
- 관련: #90
